+++ aliases.go
@@ -184,6 +184,15 @@ const ResourceRoleOrgMember = shared.ResourceRoleOrgMember
 const ResourceRoleOrgRunnersAdmin = shared.ResourceRoleOrgRunnersAdmin
 
 // This is an alias to an internal value.
+const ResourceRoleOrgProjectsAdmin = shared.ResourceRoleOrgProjectsAdmin
+
+// This is an alias to an internal value.
+const ResourceRoleOrgAutomationsAdmin = shared.ResourceRoleOrgAutomationsAdmin
+
+// This is an alias to an internal value.
+const ResourceRoleOrgGroupsAdmin = shared.ResourceRoleOrgGroupsAdmin
+
+// This is an alias to an internal value.
 const ResourceRoleGroupAdmin = shared.ResourceRoleGroupAdmin
 
 // This is an alias to an internal value.
@@ -324,6 +333,9 @@ const ResourceRoleWebhookAdmin = shared.ResourceRoleWebhookAdmin
 // This is an alias to an internal value.
 const ResourceRoleWebhookViewer = shared.ResourceRoleWebhookViewer
 
+// This is an alias to an internal value.
+const ResourceRoleWarmpoolRunner = shared.ResourceRoleWarmpoolRunner
+
 // This is an alias to an internal type.
 type ResourceType = shared.ResourceType
 
@@ -461,6 +473,18 @@ const ResourceTypeScimConfiguration = shared.ResourceTypeScimConfiguration
 
 // This is an alias to an internal value.
 const ResourceTypeServiceAccountSecret = shared.ResourceTypeServiceAccountSecret
+
+// This is an alias to an internal value.
+const ResourceTypeAnnouncementBanner = shared.ResourceTypeAnnouncementBanner
+
+// This is an alias to an internal value.
+const ResourceTypeServiceAccountToken = shared.ResourceTypeServiceAccountToken
+
+// This is an alias to an internal value.
+const ResourceTypeRoleAssignment = shared.ResourceTypeRoleAssignment
+
+// This is an alias to an internal value.
+const ResourceTypeWarmPool = shared.ResourceTypeWarmPool
 
 // This is an alias to an internal type.
 type RunsOn = shared.RunsOn